split naive.rs.
