+++ shader/ray-tracing/src/backend/wavefront_compute/geometry/naive/mod.rs
@@ -1,13 +1,21 @@
+#[cfg(test)]
+mod test;
+#[cfg(test)]
+pub(crate) use test::init_default_acceleration_structure;
+
+mod traverse_cpu;
+mod traverse_gpu;
+
 use std::ops::{BitAnd, Deref, Range};
 use std::sync::{RwLock, RwLockReadGuard};
 
 use rendiation_geometry::Box3;
 use rendiation_space_algorithm::bvh::*;
 use rendiation_space_algorithm::utils::TreeBuildOption;
+use traverse_cpu::*;
+use traverse_gpu::*;
 
-use crate::backend::wavefront_compute::geometry::{
-  intersect_ray_triangle_cpu, intersect_ray_triangle_gpu, Ray,
-};
+use crate::backend::wavefront_compute::geometry::{intersect_ray_triangle_gpu, Ray};
 use crate::*;
 
 #[repr(C)]
@@ -407,53 +415,6 @@ impl NaiveSahBvhSource {
   }
 }
 
-#[derive(Debug)]
-struct NaiveSahBvhCpu {
-  // global bvh, root at 0, content_range to index tlas_data/tlas_bounding
-  tlas_bvh_forest: Vec<DeviceBVHNode>,
-  // acceleration_structure_handle to index blas_meta_info
-  tlas_data: Vec<TopLevelAccelerationStructureSourceDeviceInstance>,
-  tlas_bounding: Vec<TlasBounding>,
-
-  // tri_range to index tri_bvh_root, box_range to index box_bvh_root
-  blas_meta_info: Vec<BlasMetaInfo>,
-  // vec3(tri_bvh_forest root_idx, geometry_idx, primitive_start, dummy)
-  tri_bvh_root: Vec<Vec4<u32>>,
-  // vec3(box_bvh_forest root_idx, geometry_idx, primitive_start, dummy)
-  box_bvh_root: Vec<Vec4<u32>>,
-  // content range to index indices
-  tri_bvh_forest: Vec<DeviceBVHNode>,
-  // content range to index boxes
-  box_bvh_forest: Vec<DeviceBVHNode>,
-
-  indices: Vec<u32>,
-  vertices: Vec<Vec3<f32>>,
-  boxes: Vec<Vec3<f32>>,
-}
-#[derive(Clone)]
-struct NaiveSahBvhGpu {
-  // global bvh, root at 0, content_range to index tlas_data/tlas_bounding
-  tlas_bvh_forest: StorageBufferReadOnlyDataView<[DeviceBVHNode]>,
-  // acceleration_structure_handle to index blas_meta_info
-  tlas_data: StorageBufferReadOnlyDataView<[TopLevelAccelerationStructureSourceDeviceInstance]>,
-  tlas_bounding: StorageBufferReadOnlyDataView<[TlasBounding]>,
-
-  // tri_range to index tri_bvh_root, box_range to index box_bvh_root
-  blas_meta_info: StorageBufferReadOnlyDataView<[BlasMetaInfo]>,
-  // vec3(tri_bvh_forest root_idx, geometry_idx, primitive_start, dummy)
-  tri_bvh_root: StorageBufferReadOnlyDataView<[Vec4<u32>]>,
-  // vec3(box_bvh_forest root_idx, geometry_idx, primitive_start, dummy)
-  box_bvh_root: StorageBufferReadOnlyDataView<[Vec4<u32>]>,
-  // content range to index indices
-  tri_bvh_forest: StorageBufferReadOnlyDataView<[DeviceBVHNode]>,
-  // content range to index boxes
-  box_bvh_forest: StorageBufferReadOnlyDataView<[DeviceBVHNode]>,
-
-  indices: StorageBufferReadOnlyDataView<[u32]>,
-  vertices: StorageBufferReadOnlyDataView<[f32]>,
-  boxes: StorageBufferReadOnlyDataView<[f32]>,
-}
-
 #[derive(Clone)]
 pub struct NaiveSahBVHSystem {
   inner: Arc<RwLock<NaiveSahBVHSystemInner>>,
@@ -503,714 +464,6 @@ impl NaiveSahBVHSystemInner {
     self
       .source
       .build(device, &mut self.cpu_data, &mut self.gpu_data);
-  }
-}
-
-impl GPUAccelerationStructureSystemCompImplInstance for NaiveSahBvhGpu {
-  fn build_shader(
-    &self,
-    compute_cx: &mut ShaderComputePipelineBuilder,
-  ) -> Box<dyn GPUAccelerationStructureSystemCompImplInvocationTraversable> {
-    let tlas_bvh_forest = compute_cx.bind_by(&self.tlas_bvh_forest);
-    let tlas_data = compute_cx.bind_by(&self.tlas_data);
-    let tlas_bounding = compute_cx.bind_by(&self.tlas_bounding);
-    let blas_meta_info = compute_cx.bind_by(&self.blas_meta_info);
-    let tri_bvh_root = compute_cx.bind_by(&self.tri_bvh_root);
-    let box_bvh_root = compute_cx.bind_by(&self.box_bvh_root);
-    let tri_bvh_forest = compute_cx.bind_by(&self.tri_bvh_forest);
-    let box_bvh_forest = compute_cx.bind_by(&self.box_bvh_forest);
-    let indices = compute_cx.bind_by(&self.indices);
-    let vertices = compute_cx.bind_by(&self.vertices);
-    let boxes = compute_cx.bind_by(&self.boxes);
-
-    let instance = NaiveSahBVHInvocationInstance {
-      tlas_bvh_forest,
-      tlas_data,
-      tlas_bounding,
-      blas_meta_info,
-      tri_bvh_root,
-      box_bvh_root,
-      tri_bvh_forest,
-      box_bvh_forest,
-      indices,
-      vertices,
-      boxes,
-    };
-
-    Box::new(instance)
-  }
-
-  fn bind_pass(&self, builder: &mut BindingBuilder) {
-    builder.bind(&self.tlas_bvh_forest);
-    builder.bind(&self.tlas_data);
-    builder.bind(&self.tlas_bounding);
-    builder.bind(&self.blas_meta_info);
-    builder.bind(&self.tri_bvh_root);
-    builder.bind(&self.box_bvh_root);
-    builder.bind(&self.tri_bvh_forest);
-    builder.bind(&self.box_bvh_forest);
-    builder.bind(&self.indices);
-    builder.bind(&self.vertices);
-    builder.bind(&self.boxes);
-  }
-}
-
-pub struct NaiveSahBVHInvocationInstance {
-  tlas_bvh_forest: ReadOnlyStorageNode<[DeviceBVHNode]>,
-  tlas_data: ReadOnlyStorageNode<[TopLevelAccelerationStructureSourceDeviceInstance]>,
-  tlas_bounding: ReadOnlyStorageNode<[TlasBounding]>,
-  blas_meta_info: ReadOnlyStorageNode<[BlasMetaInfo]>,
-  tri_bvh_root: ReadOnlyStorageNode<[Vec4<u32>]>,
-  box_bvh_root: ReadOnlyStorageNode<[Vec4<u32>]>,
-  tri_bvh_forest: ReadOnlyStorageNode<[DeviceBVHNode]>,
-  box_bvh_forest: ReadOnlyStorageNode<[DeviceBVHNode]>,
-  indices: ReadOnlyStorageNode<[u32]>,
-  vertices: ReadOnlyStorageNode<[f32]>,
-  boxes: ReadOnlyStorageNode<[f32]>,
-}
-
-struct TraverseBvhIteratorCpu<'a> {
-  bvh: &'a [DeviceBVHNode],
-  ray_origin: Vec3<f32>,
-  ray_direction: Vec3<f32>,
-  ray_range: Vec2<f32>,
-
-  curr_idx: u32,
-}
-impl<'a> Iterator for TraverseBvhIteratorCpu<'a> {
-  type Item = u32;
-  fn next(&mut self) -> Option<Self::Item> {
-    while self.curr_idx != INVALID_NEXT {
-      let node = &self.bvh[self.curr_idx as usize];
-      if intersect_ray_aabb_cpu(
-        self.ray_origin,
-        self.ray_direction,
-        self.ray_range,
-        node.aabb_min,
-        node.aabb_max,
-      ) {
-        let curr = self.curr_idx;
-        self.curr_idx = node.hit_next;
-
-        if node.hit_next == node.miss_next {
-          // is leaf
-          return Some(curr);
-        }
-      } else {
-        self.curr_idx = node.miss_next;
-      };
-    }
-
-    None
-  }
-}
-
-struct TraverseBvhIteratorGpu {
-  bvh: ReadOnlyStorageNode<[DeviceBVHNode]>,
-  ray: Node<Ray>,
-  node_idx: LocalVarNode<u32>,
-}
-impl ShaderIterator for TraverseBvhIteratorGpu {
-  type Item = Node<Vec2<u32>>; // node content range
-  fn shader_next(&self) -> (Node<bool>, Self::Item) {
-    let has_next = val(false).make_local_var();
-    let item = zeroed_val().make_local_var();
-
-    loop_by(|loop_cx| {
-      let idx = self.node_idx.load();
-      if_by(idx.equals(val(INVALID_NEXT)), || loop_cx.do_break());
-      let node = self.bvh.index(idx).load().expand();
-      let hit_aabb = intersect_ray_aabb_gpu(self.ray, node.aabb_min, node.aabb_max);
-
-      if_by(hit_aabb, || {
-        let is_leaf = node.hit_next.equals(node.miss_next);
-        self.node_idx.store(node.hit_next);
-        if_by(is_leaf, || {
-          has_next.store(val(true));
-          item.store(node.content_range);
-          loop_cx.do_break();
-        });
-      })
-      .else_by(|| {
-        self.node_idx.store(node.miss_next);
-      });
-    });
-
-    (has_next.load(), item.load())
-  }
-}
-
-/// returns iterator item = tlas_data idx
-fn traverse_tlas_gpu(
-  root: Node<u32>,
-  bvh: ReadOnlyStorageNode<[DeviceBVHNode]>,
-  tlas_bounding: ReadOnlyStorageNode<[TlasBounding]>,
-  ray: Node<Ray>,
-) -> impl ShaderIterator<Item = Node<u32>> {
-  let bvh_iter = TraverseBvhIteratorGpu {
-    bvh,
-    ray,
-    node_idx: root.make_local_var(),
-  };
-  let iter = bvh_iter.flat_map(ForRange::new);
-
-  iter.filter_map(move |tlas_idx: Node<u32>| {
-    let tlas_bounding_pack = tlas_bounding.index(tlas_idx).load();
-    let tlas_bounding = tlas_bounding_pack.expand();
-    let hit_tlas = intersect_ray_aabb_gpu(ray, tlas_bounding.world_min, tlas_bounding.world_max);
-
-    let ray = ray.expand();
-    let pass_mask = ray.mask.bitand(tlas_bounding.mask).not_equals(val(0));
-
-    // todo handle flags?
-    let hit = hit_tlas.and(pass_mask);
-
-    (hit, tlas_idx)
-  })
-}
-
-struct NaiveIntersectReporter<'a> {
-  launch_info: RayLaunchInfo,
-  world_ray: WorldRayInfo,
-  hit_ctx: HitCtxInfo,
-  closest_hit_ctx_info: &'a HitCtxInfoVar,
-  closest_hit_info: &'a HitInfoVar,
-  any_hit: &'a dyn Fn(&RayAnyHitCtx) -> Node<RayAnyHitBehavior>,
-}
-impl<'a> IntersectionReporter for NaiveIntersectReporter<'a> {
-  fn report_intersection(&self, hit_t: Node<f32>, hit_kind: Node<u32>) -> Node<bool> {
-    let r = val(false).make_local_var();
-    if_by(
-      hit_t.less_than(self.closest_hit_info.hit_distance.load()), // todo test hit_t in range
-      || {
-        let any_hit_ctx = RayAnyHitCtx {
-          launch_info: self.launch_info,
-          world_ray: self.world_ray,
-          hit_ctx: self.hit_ctx,
-          hit: HitInfo {
-            hit_kind,
-            hit_distance: hit_t,
-          },
-        };
-        let closest_hit_ctx = self.closest_hit_ctx_info;
-        let closest_hit = self.closest_hit_info;
-        let any_hit = self.any_hit;
-
-        resolve_any_hit(r, any_hit, &any_hit_ctx, closest_hit_ctx, closest_hit);
-      },
-    );
-    r.load()
-  }
-}
-
-fn resolve_any_hit(
-  r: LocalVarNode<bool>, // todo just return a node
-  any_hit: &dyn Fn(&RayAnyHitCtx) -> Node<RayAnyHitBehavior>,
-  any_hit_ctx: &RayAnyHitCtx,
-  closest_hit_ctx: &HitCtxInfoVar, // output
-  closest_hit: &HitInfoVar,        // output
-) {
-  let behavior = any_hit(any_hit_ctx);
-
-  if_by((behavior & (val(ACCEPT_HIT))).greater_than(val(0)), || {
-    // hit! update closest
-    closest_hit.test_and_store(&any_hit_ctx.hit, || {
-      closest_hit_ctx.store(&any_hit_ctx.hit_ctx)
-    });
-    // todo update ray range max
-    r.store(val(true));
-  });
-
-  if_by((behavior & val(TERMINATE_TRAVERSE)).greater_than(0), || {
-    // todo terminate
-  });
-}
-
-#[repr(C)]
-#[std430_layout]
-#[derive(ShaderStruct, Clone, Copy)]
-struct RayBlas {
-  pub ray: Ray,
-  pub blas: BlasMetaInfo,
-  pub tlas_idx: u32,
-  pub distance_scaling: f32,
-}
-
-fn iterate_tlas_blas_gpu(
-  tlas_iter: impl ShaderIterator<Item = Node<u32>>,
-  tlas_data: ReadOnlyStorageNode<[TopLevelAccelerationStructureSourceDeviceInstance]>,
-  blas_data: ReadOnlyStorageNode<[BlasMetaInfo]>,
-  ray: Node<Ray>,
-) -> impl ShaderIterator<Item = Node<RayBlas>> {
-  tlas_iter.map(move |idx: Node<u32>| {
-    let ray = ray.expand();
-    let tlas_data = tlas_data.index(idx).load().expand();
-
-    // transform ray to blas space
-    // todo check det < 0, invert cull flag?
-    let blas_ray_origin = tlas_data.transform_inv * (ray.origin, val(1.)).into();
-    let blas_ray_origin = blas_ray_origin.xyz() / blas_ray_origin.w().splat();
-    let blas_ray_direction = tlas_data.transform_inv.shrink_to_3() * ray.direction;
-    let distance_scaling = blas_ray_direction.length();
-    let blas_ray_range = ray.range * distance_scaling;
-    let blas_ray_direction = blas_ray_direction.normalize();
-
-    let blas_ray = Ray::construct(RayShaderAPIInstance {
-      origin: blas_ray_origin,
-      flags: ray.flags,
-      direction: blas_ray_direction,
-      mask: ray.mask,
-      range: blas_ray_range,
-    });
-
-    let blas_idx = tlas_data.acceleration_structure_handle;
-    let blas_data = blas_data.index(blas_idx).load();
-
-    RayBlas::construct(RayBlasShaderAPIInstance {
-      ray: blas_ray,
-      blas: blas_data,
-      tlas_idx: idx,
-      distance_scaling,
-    })
-  })
-}
-
-fn intersect_blas_gpu(
-  blas_iter: impl ShaderIterator<Item = Node<RayBlas>>,
-  tlas_data: ReadOnlyStorageNode<[TopLevelAccelerationStructureSourceDeviceInstance]>,
-  tri_bvh_root: ReadOnlyStorageNode<[Vec4<u32>]>,
-  tri_bvh_forest: ReadOnlyStorageNode<[DeviceBVHNode]>,
-  _box_bvh_root: ReadOnlyStorageNode<[Vec4<u32>]>,
-  _box_bvh_forest: ReadOnlyStorageNode<[DeviceBVHNode]>,
-  indices: ReadOnlyStorageNode<[u32]>,
-  vertices: ReadOnlyStorageNode<[f32]>,
-  _boxes: ReadOnlyStorageNode<[f32]>,
-
-  intersect: &dyn Fn(&RayIntersectCtx, &dyn IntersectionReporter),
-  any_hit: &dyn Fn(&RayAnyHitCtx) -> Node<RayAnyHitBehavior>,
-
-  launch_info: RayLaunchInfo,
-  world_ray: WorldRayInfo,
-  closest_hit_ctx_var: &HitCtxInfoVar, // output
-  closest_hit_var: &HitInfoVar,        // output
-) {
-  blas_iter.for_each(|ray_blas, _cx| {
-    let ray_blas = ray_blas.expand();
-    let ray = ray_blas.ray;
-    let blas = ray_blas.blas.expand();
-
-    ForRange::new(blas.tri_root_range).for_each(move |tri_root_idx, _cx| {
-      let geometry = tri_bvh_root.index(tri_root_idx).load();
-      let root = geometry.x();
-      let geometry_id = geometry.y();
-      let primitive_start = geometry.z();
-
-      let bvh_iter = TraverseBvhIteratorGpu {
-        bvh: tri_bvh_forest,
-        ray,
-        node_idx: root.make_local_var(),
-      };
-      let tri_idx_iter = bvh_iter.flat_map(ForRange::new); // triangle index
-
-      let ray = ray.expand();
-
-      fn read_vec3<T: ShaderNodeType>(
-        idx: Node<u32>,
-        array: ReadOnlyStorageNode<[T]>,
-      ) -> [Node<T>; 3] {
-        let i = idx * val(3);
-        let v0 = array.index(i).load();
-        let v1 = array.index(i + val(1)).load();
-        let v2 = array.index(i + val(2)).load();
-        [v0, v1, v2]
-      }
-
-      tri_idx_iter.for_each(move |tri_idx, _cx| {
-        let [i0, i1, i2] = read_vec3(tri_idx, indices);
-        let [v0x, v0y, v0z] = read_vec3(i0, vertices);
-        let [v1x, v1y, v1z] = read_vec3(i1, vertices);
-        let [v2x, v2y, v2z] = read_vec3(i2, vertices);
-        let v0 = Node::<Vec3<f32>>::from((v0x, v0y, v0z));
-        let v1 = Node::<Vec3<f32>>::from((v1x, v1y, v1z));
-        let v2 = Node::<Vec3<f32>>::from((v2x, v2y, v2z));
-        // returns (hit ? 1 : 0, distance, u, v)
-        let result = intersect_ray_triangle_gpu(ray.origin, ray.direction, ray.range, v0, v1, v2);
-        let hit = result.x().greater_than(val(0.));
-        if_by(hit, move || {
-          // todo precompute local range for intersect_ray_triangle_gpu
-          let world_distance = result.y() / ray_blas.distance_scaling;
-          // todo load tlas on every hit? protect with a bool?
-          let tlas = tlas_data.index(ray_blas.tlas_idx).load().expand();
-
-          let hit_ctx = HitCtxInfo {
-            primitive_id: tri_idx - primitive_start, // store tri offset in tri_bvh_root
-            instance_id: ray_blas.tlas_idx,
-            instance_sbt_offset: tlas.instance_shader_binding_table_record_offset,
-            instance_custom_id: tlas.instance_custom_index,
-            geometry_id,
-            object_to_world: tlas.transform_inv,
-            world_to_object: tlas.transform,
-            object_space_ray: ShaderRay {
-              origin: ray.origin,
-              direction: ray.direction,
-            },
-          };
-
-          let is_opaque = val(true); // todo check blas opaque flag
-          if_by(is_opaque, || {
-            // opaque -> invoke any_hit directly
-            let any_hit_ctx = RayAnyHitCtx {
-              launch_info,
-              world_ray,
-              hit_ctx,
-              hit: HitInfo {
-                hit_kind: val(HIT_KIND_BACK_FACING_TRIANGLE),
-                hit_distance: world_distance,
-              }, // todo specify hit local/world distance
-            };
-            let updated = val(false).make_local_var();
-            resolve_any_hit(
-              updated,
-              any_hit,
-              &any_hit_ctx,
-              closest_hit_ctx_var,
-              closest_hit_var,
-            );
-            // todo use updated?
-          })
-          .else_by(|| {
-            // non-opaque -> invode intersect
-            let intersect_ctx = RayIntersectCtx {
-              launch_info,
-              world_ray,
-              hit_ctx,
-            };
-            // intersect will invoke any_hit and then update closest_hit.
-            intersect(
-              &intersect_ctx,
-              &NaiveIntersectReporter {
-                launch_info,
-                world_ray,
-                hit_ctx,
-                closest_hit_ctx_info: closest_hit_ctx_var,
-                closest_hit_info: closest_hit_var,
-                any_hit,
-              },
-            );
-            // todo if force opaque, update intersect range to optimize
-          });
-        });
-      });
-    });
-
-    // ForRange::new(blas.box_root_range).for_each(|box_root_idx, _cx| {
-    //   let geometry = box_bvh_root.index(box_root_idx).load();
-    //   let root = geometry.x();
-    //   let geometry_id = geometry.y();
-    //
-    //   let bvh_iter = TraverseBvhIteratorGpu {
-    //     bvh: box_bvh_forest,
-    //     ray,
-    //     node_idx: root.make_local_var(),
-    //   };
-    //   let iter = bvh_iter.flat_map(ForRange::new); // box index
-    //
-    //   iter.for_each(|box_idx, _cx| {
-    //     let start = box_idx * val(2);
-    //     let min = boxes.index(indices.index(start).load()).load();
-    //     let max = boxes.index(indices.index(start + val(1)).load()).load();
-    //
-    //     let hit = intersect_ray_aabb_gpu(ray, min, max);
-    //     if_by(hit, || {
-    //       // todo call intersection with anyhit, remember distance_scaling
-    //     });
-    //   });
-    // });
-  });
-}
-
-impl NaiveSahBvhCpu {
-  fn traverse(
-    &self,
-    ray: &mut ShaderRayTraceCallStoragePayload,
-    any_hit: &mut dyn FnMut(u32, u32, f32, Vec3<f32>), /* geometry_idx, primitive_idx, distance, hit_position // todo use ctx */
-  ) {
-    // traverse tlas bvh, hit leaf
-    let tlas_iter = TraverseBvhIteratorCpu {
-      bvh: &self.tlas_bvh_forest,
-      ray_origin: ray.ray_origin,
-      ray_direction: ray.ray_direction,
-      ray_range: ray.range,
-      curr_idx: 0,
-    };
-    for hit_idx in tlas_iter {
-      let node = &self.tlas_bvh_forest[hit_idx as usize];
-
-      // for each tlas, visit blas
-      for tlas_idx in node.content_range.x..node.content_range.y {
-        // test tlas bounding
-        let tlas_bounding = &self.tlas_bounding[tlas_idx as usize];
-        if !intersect_ray_aabb_cpu(
-          ray.ray_origin,
-          ray.ray_direction,
-          ray.range,
-          tlas_bounding.world_min,
-          tlas_bounding.world_max,
-        ) {
-          continue;
-        }
-        if ray.cull_mask & tlas_bounding.mask == 0 {
-          continue;
-        }
-
-        let tlas_data = &self.tlas_data[tlas_idx as usize];
-        // hit tlas
-        let blas_idx = tlas_data.acceleration_structure_handle;
-
-        // traverse blas bvh
-        let blas_ray_origin = tlas_data.transform_inv * ray.ray_origin.expand_with_one();
-        let blas_ray_origin = blas_ray_origin.xyz() / blas_ray_origin.w();
-        let blas_ray_direction = tlas_data.transform_inv.to_mat3() * ray.ray_direction;
-        let distance_scaling = blas_ray_direction.length();
-        let blas_ray_range = ray.range * distance_scaling;
-        let blas_ray_direction = blas_ray_direction.normalize();
-
-        // todo check triangle related flags
-        let blas_meta_info = &self.blas_meta_info[blas_idx as usize];
-        for tri_root_index in blas_meta_info.tri_root_range.x..blas_meta_info.tri_root_range.y {
-          let geometry = self.tri_bvh_root[tri_root_index as usize];
-          let blas_root_idx = geometry.x;
-          let geometry_idx = geometry.y;
-          let primitive_start = geometry.z;
-
-          let bvh_iter = TraverseBvhIteratorCpu {
-            bvh: &self.tri_bvh_forest,
-            ray_origin: blas_ray_origin,
-            ray_direction: blas_ray_direction,
-            ray_range: blas_ray_range,
-            curr_idx: blas_root_idx,
-          };
-
-          for hit_idx in bvh_iter {
-            let node = &self.tri_bvh_forest[hit_idx as usize];
-
-            for tri_idx in node.content_range.x..node.content_range.y {
-              let i0 = self.indices[tri_idx as usize * 3];
-              let i1 = self.indices[tri_idx as usize * 3 + 1];
-              let i2 = self.indices[tri_idx as usize * 3 + 2];
-              let v0 = self.vertices[i0 as usize];
-              let v1 = self.vertices[i1 as usize];
-              let v2 = self.vertices[i2 as usize];
-
-              // vec4(hit, distance, u, v)
-              let intersection = intersect_ray_triangle_cpu(
-                blas_ray_origin,
-                blas_ray_direction,
-                blas_ray_range,
-                v0,
-                v1,
-                v2,
-                // todo check flags
-              );
-
-              if intersection[0] > 0. {
-                let distance = intersection[1] / distance_scaling;
-                let p = blas_ray_origin + distance * blas_ray_direction;
-                // println!("hit {p:?}");
-                let primitive_idx = tri_idx - primitive_start;
-                any_hit(geometry_idx, primitive_idx, distance, p);
-              }
-            }
-          }
-        }
-
-        // todo check box related flags
-        for box_root_index in blas_meta_info.box_root_range.x..blas_meta_info.box_root_range.y {
-          let idx = self.box_bvh_root[box_root_index as usize];
-          let blas_root_idx = idx.x;
-          // let geometry_idx = idx.y;
-
-          let box_iter = TraverseBvhIteratorCpu {
-            bvh: &self.box_bvh_forest,
-            ray_origin: blas_ray_origin,
-            ray_direction: blas_ray_direction,
-            ray_range: blas_ray_range,
-            curr_idx: blas_root_idx,
-          };
-
-          for hit_idx in box_iter {
-            let node = &self.box_bvh_forest[hit_idx as usize];
-            let aabb =
-              &self.boxes[node.content_range.x as usize * 2..node.content_range.y as usize * 2];
-            for aabb in aabb.chunks_exact(2) {
-              let hit = intersect_ray_aabb_cpu(
-                blas_ray_origin,
-                blas_ray_direction,
-                blas_ray_range,
-                aabb[0],
-                aabb[1],
-              );
-              if hit {
-                // todo call intersect, then anyhit
-                // todo modify range after hit
-              }
-            }
-          }
-        }
-      }
-    }
-  }
-}
-
-#[derive(Copy, Clone)]
-struct HitCtxInfoVar {
-  pub primitive_id: LocalVarNode<u32>,
-  pub instance_id: LocalVarNode<u32>,
-  pub instance_sbt_offset: LocalVarNode<u32>,
-  pub instance_custom_id: LocalVarNode<u32>,
-  pub geometry_id: LocalVarNode<u32>,
-  pub object_to_world: LocalVarNode<Mat4<f32>>,
-  pub world_to_object: LocalVarNode<Mat4<f32>>,
-  pub object_space_ray_origin: LocalVarNode<Vec3<f32>>,
-  pub object_space_ray_direction: LocalVarNode<Vec3<f32>>,
-}
-impl HitCtxInfoVar {
-  fn store(&self, source: &HitCtxInfo) {
-    self.primitive_id.store(source.primitive_id);
-    self.instance_id.store(source.instance_id);
-    self.instance_sbt_offset.store(source.instance_sbt_offset);
-    self.instance_custom_id.store(source.instance_custom_id);
-    self.geometry_id.store(source.geometry_id);
-    self.object_to_world.store(source.object_to_world);
-    self.world_to_object.store(source.world_to_object);
-    self
-      .object_space_ray_origin
-      .store(source.object_space_ray.origin);
-    self
-      .object_space_ray_direction
-      .store(source.object_space_ray.direction);
-  }
-}
-#[derive(Copy, Clone)]
-struct HitInfoVar {
-  pub any_hit: LocalVarNode<bool>,
-  pub hit_kind: LocalVarNode<u32>,
-  pub hit_distance: LocalVarNode<f32>,
-}
-impl HitInfoVar {
-  fn test_and_store(&self, source: &HitInfo, if_passed: impl FnOnce()) {
-    if_by(
-      source.hit_distance.less_than(self.hit_distance.load()),
-      || {
-        self.any_hit.store(val(true));
-        self.hit_kind.store(source.hit_kind);
-        self.hit_distance.store(source.hit_distance);
-        if_passed();
-      },
-    );
-  }
-}
-
-impl GPUAccelerationStructureSystemCompImplInvocationTraversable for NaiveSahBVHInvocationInstance {
-  fn traverse(
-    &self,
-    // todo launch info
-    trace_payload: ENode<ShaderRayTraceCallStoragePayload>,
-    intersect: &dyn Fn(&RayIntersectCtx, &dyn IntersectionReporter),
-    any_hit: &dyn Fn(&RayAnyHitCtx) -> Node<RayAnyHitBehavior>,
-  ) -> ShaderOption<RayClosestHitCtx> {
-    let ray = Ray::construct(RayShaderAPIInstance {
-      origin: trace_payload.ray_origin,
-      flags: trace_payload.ray_flags,
-      direction: trace_payload.ray_direction,
-      mask: trace_payload.cull_mask,
-      range: trace_payload.range,
-    });
-
-    let tlas_idx_iter = traverse_tlas_gpu(val(0), self.tlas_bvh_forest, self.tlas_bounding, ray);
-
-    let blas_iter = iterate_tlas_blas_gpu(tlas_idx_iter, self.tlas_data, self.blas_meta_info, ray);
-
-    // construct ctx
-    let launch_info = RayLaunchInfo {
-      launch_id: val(vec3(0, 0, 0)),   // todo
-      launch_size: val(vec3(0, 0, 0)), // todo
-    };
-    let world_ray = WorldRayInfo {
-      world_ray: ShaderRay {
-        origin: trace_payload.ray_origin,
-        direction: trace_payload.ray_direction,
-      },
-      ray_range: ShaderRayRange {
-        min: trace_payload.range.x(),
-        max: trace_payload.range.y(),
-      },
-      ray_flags: trace_payload.ray_flags,
-    };
-
-    let hit_ctx_info_var = HitCtxInfoVar {
-      primitive_id: val(0).make_local_var(),
-      instance_id: val(0).make_local_var(),
-      instance_sbt_offset: val(0).make_local_var(),
-      instance_custom_id: val(0).make_local_var(),
-      geometry_id: val(0).make_local_var(),
-      object_to_world: mat4_identity_node().make_local_var(),
-      world_to_object: mat4_identity_node().make_local_var(),
-      object_space_ray_origin: val(vec3(0., 0., 0.)).make_local_var(),
-      object_space_ray_direction: val(vec3(0., 0., 0.)).make_local_var(),
-    };
-
-    let hit_info_var = HitInfoVar {
-      any_hit: val(false).make_local_var(),
-      hit_kind: val(0).make_local_var(),
-      hit_distance: world_ray.ray_range.max.make_local_var(),
-    };
-
-    intersect_blas_gpu(
-      blas_iter,
-      self.tlas_data,
-      self.tri_bvh_root,
-      self.tri_bvh_forest,
-      self.box_bvh_root,
-      self.box_bvh_forest,
-      self.indices,
-      self.vertices,
-      self.boxes,
-      intersect,
-      any_hit,
-      launch_info,
-      world_ray,
-      &hit_ctx_info_var, // output
-      &hit_info_var,     // output
-    );
-
-    let hit_ctx_info = HitCtxInfo {
-      primitive_id: hit_ctx_info_var.primitive_id.load(),
-      instance_id: hit_ctx_info_var.instance_id.load(),
-      instance_sbt_offset: hit_ctx_info_var.instance_sbt_offset.load(),
-      instance_custom_id: hit_ctx_info_var.instance_custom_id.load(),
-      geometry_id: hit_ctx_info_var.geometry_id.load(),
-      object_to_world: hit_ctx_info_var.object_to_world.load(),
-      world_to_object: hit_ctx_info_var.world_to_object.load(),
-      object_space_ray: ShaderRay {
-        origin: hit_ctx_info_var.object_space_ray_origin.load(),
-        direction: hit_ctx_info_var.object_space_ray_direction.load(),
-      },
-    };
-    let hit_info = HitInfo {
-      hit_kind: hit_info_var.hit_kind.load(),
-      hit_distance: hit_info_var.hit_distance.load(),
-    };
-
-    ShaderOption {
-      is_some: hit_info_var.any_hit.load(),
-      payload: RayClosestHitCtx {
-        launch_info,
-        world_ray,
-        hit_ctx: hit_ctx_info,
-        hit: hit_info,
-      },
-    }
   }
 }
 
@@ -1278,289 +531,4 @@ fn compute_bvh_next<B: BVHBounding>(flatten_nodes: &[FlattenBVHNode<B>]) -> Vec<
     result.push((hit, miss));
   }
   result
-}
-
-#[cfg(test)]
-pub(crate) fn init_default_acceleration_structure(
-  system: &dyn GPUAccelerationStructureSystemProvider,
-) {
-  #[rustfmt::skip]
-  const CUBE_POSITION: [f32; 72] = [
-     0.5,  0.5,  0.5, -0.5,  0.5,  0.5, -0.5, -0.5,  0.5,  0.5, -0.5,  0.5, // v0,v1,v2,v3 (front)
-     0.5,  0.5,  0.5,  0.5, -0.5,  0.5,  0.5, -0.5, -0.5,  0.5,  0.5, -0.5, // v0,v3,v4,v5 (right)
-     0.5,  0.5,  0.5,  0.5,  0.5, -0.5, -0.5,  0.5, -0.5, -0.5,  0.5,  0.5, // v0,v5,v6,v1 (top)
-    -0.5,  0.5,  0.5, -0.5,  0.5, -0.5, -0.5, -0.5, -0.5, -0.5, -0.5,  0.5, // v1,v6,v7,v2 (left)
-    -0.5, -0.5, -0.5,  0.5, -0.5, -0.5,  0.5, -0.5,  0.5, -0.5, -0.5,  0.5, // v7,v4,v3,v2 (bottom)
-     0.5, -0.5, -0.5, -0.5, -0.5, -0.5, -0.5,  0.5, -0.5,  0.5,  0.5, -0.5, // v4,v7,v6,v5 (back)
-  ];
-  #[rustfmt::skip]
-  const CUBE_INDEX: [u16; 36] = [
-     0, 1, 2,   2, 3, 0,    // v0-v1-v2, v2-v3-v0 (front)
-     4, 5, 6,   6, 7, 4,    // v0-v3-v4, v4-v5-v0 (right)
-     8, 9,10,  10,11, 8,    // v0-v5-v6, v6-v1-v0 (top)
-    12,13,14,  14,15,12,    // v1-v6-v7, v7-v2-v1 (left)
-    16,17,18,  18,19,16,    // v7-v4-v3, v3-v2-v7 (bottom)
-    20,21,22,  22,23,20,    // v4-v7-v6, v6-v5-v4 (back)
-  ];
-
-  let blas_handle = system.create_bottom_level_acceleration_structure(&[
-    BottomLevelAccelerationStructureBuildSource::Triangles {
-      positions: CUBE_POSITION
-        .chunks_exact(3)
-        .map(|abc| vec3(abc[0], abc[1], abc[2]))
-        .collect(),
-      indices: CUBE_INDEX.map(|i| i as u32).into_iter().collect(),
-    },
-  ]);
-
-  fn add_tlas(
-    system: &dyn GPUAccelerationStructureSystemProvider,
-    transform: Mat4<f32>,
-    blas_handle: &BottomLevelAccelerationStructureHandle,
-  ) -> Box<dyn GPUAccelerationStructureInstanceProvider> {
-    system.create_top_level_acceleration_structure(&[TopLevelAccelerationStructureSourceInstance {
-      transform,
-      instance_custom_index: 0,
-      mask: u32::MAX,
-      instance_shader_binding_table_record_offset: 0,
-      flags: 0,
-      acceleration_structure_handle: *blas_handle,
-    }])
-  }
-  for i in -2..=2 {
-    for j in -2..=2 {
-      add_tlas(
-        system,
-        Mat4::translate((i as f32 * 1.5, j as f32 * 1.5, -10.)),
-        &blas_handle,
-      );
-    }
-  }
-  add_tlas(
-    system,
-    Mat4::translate((0., 4.5, -10.)) * Mat4::scale((5., 1., 1.)),
-    &blas_handle,
-  );
-  add_tlas(
-    system,
-    Mat4::translate((0., -4.5, -10.))
-      * Mat4::rotate_y(std::f32::consts::PI)
-      * Mat4::scale((5., 1., 1.)),
-    &blas_handle,
-  );
-  add_tlas(
-    system,
-    Mat4::translate((4.5, -4.5, -10.))
-      * Mat4::rotate_y(std::f32::consts::PI * 0.5)
-      * Mat4::scale((5., 1., 1.)),
-    &blas_handle,
-  );
-  add_tlas(
-    system,
-    Mat4::translate((-4.5, -4.5, -10.))
-      * Mat4::rotate_y(std::f32::consts::PI * -0.5)
-      * Mat4::scale((5., 1., 1.)),
-    &blas_handle,
-  );
-}
-
-#[test]
-fn test_both_triangle() {
-  test_gpu_triangle();
-  test_cpu_triangle();
-}
-
-#[test]
-fn test_cpu_triangle() {
-  const W: usize = 256;
-  const H: usize = 256;
-  const FAR: f32 = 100.;
-  const ORIGIN: Vec3<f32> = vec3(0., 0., 0.);
-  // const GEOMETRY_IDX_MAX: u32 = 1;
-  const PRIMITIVE_IDX_MAX: u32 = 12;
-
-  let (gpu, _) = futures::executor::block_on(GPU::new(Default::default())).unwrap();
-  let system = NaiveSahBVHSystem::new(gpu);
-  init_default_acceleration_structure(&system);
-
-  let _ = system.get_or_build_gpu_data(); // trigger build
-  let inner = system.inner.read().unwrap();
-  let cpu_data = inner.cpu_data.as_ref().unwrap();
-
-  let mut payload = ShaderRayTraceCallStoragePayload::zeroed();
-  payload.cull_mask = u32::MAX;
-  payload.range = vec2(0., FAR);
-
-  let mut out = Box::new([[(FAR, 0); W]; H]);
-
-  for j in 0..H {
-    for i in 0..W {
-      let x = (i as f32 + 0.5) / W as f32 * 2. - 1.;
-      let y = 1. - (j as f32 + 0.5) / H as f32 * 2.;
-      let target = vec3(x, y, -1.); // fov = 90 deg
-      let direction = (target - ORIGIN).normalize();
-
-      payload.ray_origin = ORIGIN;
-      payload.ray_direction = direction;
-      cpu_data.traverse(
-        &mut payload,
-        &mut |_geometry_id, primitive_id, distance, _position| {
-          let (d, id) = &mut out[j][i];
-          if distance < *d {
-            *d = distance;
-            *id = primitive_id + 1;
-          }
-        },
-      );
-    }
-  }
-
-  let mut file = format!("P2\n{W} {H}\n{PRIMITIVE_IDX_MAX}\n");
-  for j in 0..H {
-    file.push_str(out[j].map(|(_, id)| format!("{id}")).join(" ").as_str());
-    file.push('\n');
-  }
-  std::fs::write("trace_cpu.pbm", file).unwrap();
-}
-
-#[test]
-fn test_gpu_triangle() {
-  const H: usize = 256;
-  const W: usize = 256;
-  const FAR: f32 = 100.;
-  const ORIGIN: Vec3<f32> = vec3(0., 0., 0.);
-  // const GEOMETRY_IDX_MAX: u32 = 1;
-  const PRIMITIVE_IDX_MAX: u32 = 12;
-
-  let mut direction = vec![vec4(0., 0., 0., 0.); H * W];
-  for j in 0..H {
-    for i in 0..W {
-      let x = (i as f32 + 0.5) / W as f32 * 2. - 1.;
-      let y = 1. - (j as f32 + 0.5) / H as f32 * 2.;
-      let target = vec3(x, y, -1.); // fov = 90 deg
-      let dir = (target - ORIGIN).normalize();
-      direction[j * W + i] = vec4(dir.x, dir.y, dir.z, 0.);
-    }
-  }
-
-  let (gpu, _) = futures::executor::block_on(GPU::new(Default::default())).unwrap();
-  let mut cx = DeviceParallelComputeCtx::new(&gpu);
-
-  let direction = Box::new(direction) as Box<dyn DeviceParallelCompute<Node<Vec4<f32>>>>;
-  let tester = GpuTester::new(direction, gpu);
-
-  cx.force_indirect_dispatch = false;
-  let (_, _size, result) = futures::executor::block_on(tester.read_back_host(&mut cx)).unwrap();
-  // println!("result {:?} {:?}", result.len(), result);
-
-  let mut file = format!("P2\n{W} {H}\n{PRIMITIVE_IDX_MAX}\n");
-  for j in 0..H {
-    file.push_str(
-      result[j * W..(j + 1) * W]
-        .iter()
-        .map(|v| format!("{v}"))
-        .collect::<Vec<_>>()
-        .join(" ")
-        .as_str(),
-    );
-    file.push('\n');
-  }
-  std::fs::write("trace_gpu.pbm", file).unwrap();
-
-  #[derive(Clone)]
-  struct GpuTester {
-    upstream: Box<dyn DeviceParallelCompute<Node<Vec4<f32>>>>, // ray direction
-    system: NaiveSahBVHSystem,
-  }
-  struct GpuTesterInner {
-    upstream: Box<dyn DeviceInvocationComponent<Node<Vec4<f32>>>>, // ray direction
-    system: NaiveSahBvhGpu,
-  }
-
-  impl GpuTester {
-    fn new(upstream: Box<dyn DeviceParallelCompute<Node<Vec4<f32>>>>, gpu: GPU) -> Self {
-      let system = NaiveSahBVHSystem::new(gpu);
-      init_default_acceleration_structure(&system);
-      Self { upstream, system }
-    }
-  }
-  impl DeviceParallelCompute<Node<u32>> for GpuTester {
-    fn execute_and_expose(
-      &self,
-      cx: &mut DeviceParallelComputeCtx,
-    ) -> Box<dyn DeviceInvocationComponent<Node<u32>>> {
-      Box::new(GpuTesterInner {
-        upstream: self.upstream.execute_and_expose(cx),
-        system: self.system.get_or_build_gpu_data().clone(),
-      })
-    }
-    fn result_size(&self) -> u32 {
-      self.upstream.result_size()
-    }
-  }
-  impl DeviceParallelComputeIO<u32> for GpuTester {}
-
-  impl ShaderHashProvider for GpuTesterInner {
-    fn hash_pipeline(&self, hasher: &mut PipelineHasher) {
-      self.upstream.hash_pipeline_with_type_info(hasher)
-    }
-    shader_hash_type_id! {}
-  }
-  impl DeviceInvocationComponent<Node<u32>> for GpuTesterInner {
-    fn work_size(&self) -> Option<u32> {
-      self.upstream.work_size()
-    }
-    fn build_shader(
-      &self,
-      builder: &mut ShaderComputePipelineBuilder,
-    ) -> Box<dyn DeviceInvocation<Node<u32>>> {
-      builder.log_result = true;
-
-      let upstream_shader = self.upstream.build_shader(builder);
-
-      let traversable = self.system.clone().build_shader(builder);
-
-      upstream_shader
-        .adhoc_invoke_with_self_size(move |upstream, id| {
-          let (input, valid) = upstream.invocation_logic(id);
-
-          let payload = ShaderRayTraceCallStoragePayloadShaderAPIInstance {
-            payload_ref: val(0),
-            tlas_idx: val(0), // todo support tlas selection
-            ray_flags: val(0),
-            cull_mask: val(u32::MAX),
-            sbt_ray_config_offset: val(0),
-            sbt_ray_config_stride: val(0),
-            miss_index: val(0),
-            ray_origin: val(ORIGIN),
-            ray_direction: input.xyz(),
-            range: val(vec2(0., FAR)),
-          };
-
-          // todo how to access user payload?
-          let output =
-            traversable.traverse(payload, &|_ctx, _reporter| {}, &|_ctx| val(ACCEPT_HIT));
-          (
-            output.payload.hit_ctx.primitive_id + output.is_some.into_u32(),
-            valid,
-          )
-        })
-        .into_boxed()
-    }
-    fn bind_input(&self, builder: &mut BindingBuilder) {
-      self.upstream.bind_input(builder);
-      self.system.bind_pass(builder);
-    }
-    fn requested_workgroup_size(&self) -> Option<u32> {
-      self.upstream.requested_workgroup_size()
-    }
-  }
-}
-
-fn mat4_identity_node() -> Node<Mat4<f32>> {
-  let x = val(vec4(1., 0., 0., 0.));
-  let y = val(vec4(0., 1., 0., 0.));
-  let z = val(vec4(0., 0., 1., 0.));
-  let w = val(vec4(0., 0., 0., 1.));
-  (x, y, z, w).into()
 }
